3.14 lightweight frame compatibliity
Summary:
This fixes up our lightweight frame support so that it builds on 3.14 when running on top of D87386838 which is a similar implementation to our 3.12 implementation.

We still need the support in the 3.14 runtime and that will be different than what we have in 3.12 - it will likely leverage the [work upstream](https://github.com/python/cpython/issues/100987) to turn `f_code` into `f_executable` and add a new executor type. That'll necessitate other changes to how lightweight frames work in CinderX on 3.14 but this at least gets us to a place where we know the baseline implementation is working similar to 3.12.

Reviewed By: alexmalyshev

Differential Revision: D87386840

fbshipit-source-id: 0941eddbb7e47d99eaa470b3aa20be0d4442eb3e