Add HIR stats change test
Summary:
The idea is to flag changes to generated HIR that may not affect correctness, but could be indicative of optimization regressions.

Builds on top of the HIR stats summary used to identify some regressions during the 3.12 JIT upgrade. A "workload" set of tests to generate some sample HIR stats and then checks to see if these have change against the base revision. If there's a change the test fails. As this might happen for legitimate reasons the test can be suppressed on Sandcastle runs by adding a `no-hir-stats` tag to the diff.

The workload set of tests is mostly generated from `python3.14 -mtest --pgo --list-tests` but with a bit of pruning to take out tests that don't work well in this setup. As DinoV points out this could be a bit fragile if there are changes to tests in CPython. We can address this by having two versions of this tests which mutually exclusively run CinderX or CPython tests and trigger based on changes to CPython or CinderX respectively. For now, hopefully this will be infrequent enough that the suppression tag can be used to dismiss noise without it becoming too bothersome.

Reviewed By: DinoV

Differential Revision: D88819156

fbshipit-source-id: 0510f169d28eb7ae54df2568680218b0bd01a4d4