Store + dump debug stats about JIT compiled functions
Summary:
Adds a new `-X jit-dump-hir-stats` options which writes out JSON formatted data in the JIT log per compile function, containing the number of instruction types, and type types. There is also a script for parsing the log data and producing a condensed report.

This was useful for spotting regressions between 3.10 and 3.12 generated code.

The parsing script I must admit is fully AI generated and I've barely looked at it.

Reviewed By: DinoV

Differential Revision: D78605511

fbshipit-source-id: f04894ddfac39c531278318d52fd0f8883c95c3f