Back out "Revert D88974414: Back out "Revert D87386840: [CinderX][CX314] 3.14 lightweight frame compatibliity""
Summary:
Original commit changeset: 379320f6b749

Original Phabricator Diff: D88974414

Take 3, this time with feeling...

This is the same as the previous version but we no longer do the `jitFrameSetFunction(frame, nullptr);` in `jitFrameClearExceptCode`. This was necessary when developing the 3.14 "hybrid" version that worked on a forward port of our 3.12 implementation. In 3.12 the function is a stack ref, and in debug builds there is some extra verification that happens on assignment. During JIT generator deopt and finalization we can copy a dangling pointer from the FrameHeader into the `_PyInterpreterFrame` and this would blow up in the debug builds. So clearing it would avoid this, but also apparently cause some crashes in IG prod.

But that's no longer necessary in the new 3.14 implementation - in fact in v1-v2 of D88432115 the handling of the non-rtfs function moves under a `if constexpr (PY_VERSION_HEX < 0x030E0000) {`.

So this just deletes the unnecessary complication.

Reviewed By: alexmalyshev

Differential Revision: D89083281

fbshipit-source-id: 9b2587cc5f6456a7650693e7c50762caac23889c