Support lightweight frames in CinderX
Summary:
Adds support for lightweight frames to CinderX 3.14. We use the JIT reifier and create an instance per hir Function that flows down to the compiled code (this needs to be done at pre-load time before we've gone multi-threaded).

The reifier is stored in the `f_executable` instead of `f_funcobj`.

For normal Python functions we no longer store the function object in the `FrameHeader` - there's no need to as `f_funcobj` is available to us. We do store the `RuntimeFrameState` for inlined functions as we do still need to be able to get back to it.

Reviewed By: alexmalyshev

Differential Revision: D88432115

fbshipit-source-id: 178407d5650b86f70639fed82b9d0a9eff775422