Make it clear we only link lightweight frames for generators
Summary: We have this check to see if we're a generator frame but as far as the actual CinderX runtime is concerned we only call this for generator frames as JITed functions init their own lightweight frames on the stack. We do use this in some tests as well so we can't assert on it, but we don't need the runtime check for any reason.

Reviewed By: alexmalyshev

Differential Revision: D88213599

fbshipit-source-id: 552b9dd57ecf61b25aa233fc1ce71430be9f93e4